refactor(web): single runClientAction bridge between typed API clients and runAction (W01)
